==> Reg1016/1-NewDataReg.ipynb <==
import scipy as sp
import numpy as np
import pandas as pd
import timeit
import re
import json
import pickle
import fastparquet
import os
os.chdir('/mnt/t48/bighomes-active/sfeng/patentdiffusion/')
seed = 3
import seaborn as sns
import matplotlib
import matplotlib.pyplot as plt
from sklearn.preprocessing import StandardScaler
scaler = StandardScaler()
import datetime
import statsmodels.api as sm
import statsmodels.formula.api as smf
import statsmodels.iolib.summary2 as summary2
import logging

Code based from:
- https://sfengc7.stern.nyu.edu:8888/notebooks/patentdiffusion/201808Results/Reg0828/2c-AddingControls-0828.ipynb

## Assign each reference text a unique identifier
- The "uuid" in the `otherreference.tsv` files is unique for every single reference
- Add unique identifier using haslib

oc = pd.read_table("RawData/PatentsView/otherreference.tsv")
# Remove missing
oc = oc.dropna(how="any")

import hashlib
%time oc["ref_id"] = oc["text"].apply(lambda x: hashlib.sha1(str.encode(x)).hexdigest())

# Check it worked
print(len(oc.loc[oc["ref_id"].duplicated()]))

oc = oc[["ref_id", "patent_id"]]
oc["patent_id"] = oc["patent_id"].apply(lambda x: int(x) if x.isdigit() else np.nan)
oc = oc.loc[oc["patent_id"].notnull()]
oc["patent_id"] = oc["patent_id"].astype(int)

fastparquet.write("RawData/Cleaned/otherreference1016.parq", oc, compression="GZIP")

# %time oc = fastparquet.ParquetFile("RawData/Cleaned/otherreference1016.parq").to_pandas()
%time oc = {n:g["ref_id"].tolist() for n,g in oc.groupby("patent_id")}

Add to all pair similarity files

files = ["DataStore/2018-10/Reg0930/naics_name_all_1004.parq",
        "DataStore/2018-10/Reg0930/primclass_all_1004.parq",
        "DataStore/2018-06/newterm_pairs_sim_0620.parq",
        "DataStore/2018-08/newterms_pats_sim_cite_0903.parq",
        "DataStore/2018-07-P2/mutual_cited_0716.parq"]

for f in files:    
    rs = fastparquet.ParquetFile(f).to_pandas()
    
    if f == "DataStore/2018-06/newterm_pairs_sim_0620.parq":
        %time num_common_cites = (set(oc.get(tp, [])).intersection(set(oc.get(op, []))) for tp, op in zip(rs["lead_pat"], rs["follow_pat"]))
    else:
        %time num_common_cites = (set(oc.get(tp, [])).intersection(set(oc.get(op, []))) for tp, op in zip(rs["tp"], rs["op"]))
    
    %time rs["num_common_npc"] = [len(i) for i in num_common_cites]
    
    # Normalize
    if f in ["DataStore/2018-10/Reg0930/naics_name_all_1004.parq",
        "DataStore/2018-10/Reg0930/primclass_all_1004.parq"]:
        rs["norm_num_common_npc"] = np.nan
        rs.loc[rs["num_common_npc"].notnull(), "norm_num_common_npc"] = \
        scaler.fit_transform(rs.loc[rs["num_common_npc"].notnull(), "num_common_npc"].values.reshape(-1,1))
    fastparquet.write(f, rs, compression="GZIP")
    del(rs)

## Patent examiners

### Clean examiners data

# re = pd.read_table("RawData/PatentsView/rawexaminer.tsv")

re = re[["patent_id", "name_first", "name_last"]].dropna(how="any")

# Create id for each examiner
re["examiner"] = [f+" "+l for f,l in zip(re["name_first"], re["name_last"])]

import hashlib
%time re["examiner_id"] = re["examiner"].apply(lambda x: hashlib.sha1(str.encode(x)).hexdigest())

re["patent_id"] = re["patent_id"].apply(lambda x: int(x) if x.isdigit() else np.nan)
re = re.loc[re["patent_id"].notnull()]
re["patent_id"] = re["patent_id"].astype(int)

print(len(re["examiner_id"].unique()))

fastparquet.write("RawData/Cleaned/patexaminer1016.parq", re, compression="GZIP")

### Create dictionary of top names for fixed effects
- Examiners
- Lawyers
- MSAs
- Primary Classes

topn = {}
topn["examiner"] = re["examiner_id"].value_counts()[:100].index.tolist()

# Lawyers
ld = pd.read_csv("RawData/Cleaned/patent_lawyer.csv", index_col=0).drop_duplicates("patent")

topn["lawyer"] = ld["lawyer_id"].value_counts()[:100].index.tolist()

# MSAs & Primary classes
pdf = fastparquet.ParquetFile("RawData/Cleaned/patent_loc_unique_us_0628.parq").to_pandas(["patent", "primclass", "inv_msa"])
dup_pats = pd.read_pickle("RawData/Cleaned/duplicate_pattext_0712.pkl").tolist()
# Get relevant US Patents
pdf = pdf.loc[~pdf["patent"].isin(dup_pats)]


topn["inv_msa"] = pdf["inv_msa"].value_counts()[:100].index.tolist()
topn["primclass"] = pdf["primclass"].value_counts()[:100].index.tolist()
del(pdf)

# Dictionary for lawyer and examiners
re = dict(zip(re["patent_id"], re["examiner_id"]))
ld = dict(zip(ld["patent"], ld["lawyer_id"]))

mod_keys = ["naics_name", "primclass"]

for k in mod_keys:    
    rs = fastparquet.ParquetFile("DataStore/2018-10/Reg0930/{0}_all_1004.parq".format(k)).to_pandas()
    
    rs["tp_examiner"] = rs["tp"].map(re)
    rs["tp_lawyer"] = rs["tp"].map(ld)
    
    # Fixed effects
    for c in ["primclass", "inv_msa", "examiner", "lawyer"]:
        rs["tp_{0}_FE".format(c)] = rs["tp_{0}".format(c)].astype(str)
        rs.loc[~(rs["tp_{0}".format(c)].isin(topn[c])), "tp_{0}_FE".format(c)] = "other"
        
        print(rs["tp_{0}_FE".format(c)].value_counts()[:10])
        # Drop original
        rs = rs.drop("tp_{0}".format(c), 1)
        
    # New examiner match
    rs["examiner_match"] = (rs["tp"].map(re) == rs["op"].map(re))
    print("Examiner match")
    print(rs["examiner_match"].value_counts())
    
    print(rs.columns)
    
    # Common cited match
    rs["common_cited_match"] = (rs["num_common_cited"] >= 1)
    rs["common_npc_match"] = (rs["num_common_npc"] >= 1)
    print("Common cited")
    print(rs["common_cited_match"].value_counts())
    print("NPC")
    print(rs["common_npc_match"].value_counts())
    
    # Common establishment inventor a lot of missing values, replace with 0
    rs.loc[rs["common_est_inv"].isnull(), "common_est_inv"] = 0
    
    # Select columns
    rs = rs[['tp', 'op', 'sim_docvecs', 'sim_ldavecs', 'tp_gyear', 'tp_naics_name',
       'op_naics_name', 'op_primclass', 'op_inv_msa', 'inv_msa_match',
       'primclass_match', 'norm_sim_ldavecs', 'norm_sim_docvecs',
       'year_group', 'common_est_inv', 'common_pat_inv', 'lawyer_match',
       'num_common_cited', 'norm_num_common_cited', 'tp_pct_common_cited',
       'norm_tp_pct_common_cited', 'common_cited_match', "common_npc_match", 'mean_sim_docvecs_pc',
       'mean_sim_ldavecs_pc', 'mean_sim_docvecs_pc_msa',
       'mean_sim_ldavecs_pc_msa', 'norm_mean_sim_docvecs_pc',
       'norm_mean_sim_ldavecs_pc', 'norm_mean_sim_docvecs_pc_msa',
       'norm_mean_sim_ldavecs_pc_msa', 'sd_sim_docvecs_pc',
       'sd_sim_ldavecs_pc', 'sd_sim_docvecs_pc_msa', 'sd_sim_ldavecs_pc_msa',
       'pc_msa_greater_0', 'pc_msa_less_0', 'num_common_npc',
       'norm_num_common_npc', 'tp_primclass_FE', 'tp_inv_msa_FE',
       'tp_examiner_FE', 'tp_lawyer_FE', 'examiner_match']]

    # Save
    fastparquet.write("DataStore/2018-10/Reg1016/{0}_all_1016.parq".format(k), rs, compression="GZIP")
        
    
    

rs.info()

### Add data for JTH Citations sample

rs = pd.read_pickle("DataStore/2018-07-P3/JTHReg0727/targ_cite_sim_reg_0727.pkl")
rs.columns

rs = pd.read_pickle("DataStore/2018-07-P3/JTHReg0727/targ_cite_sim_reg_0727.pkl")

# Patent examiner
re = fastparquet.ParquetFile("RawData/Cleaned/patexaminer1016.parq").to_pandas()

# Patent lawyers
ld = pd.read_csv("RawData/Cleaned/patent_lawyer.csv", index_col=0).drop_duplicates("patent")

# Top values
topn = {}
topn["examiner"] = re["examiner_id"].value_counts()[:100].index.tolist()
topn["lawyer"] = ld["lawyer_id"].value_counts()[:100].index.tolist()

# Use only what's in data
re = re.loc[re["patent_id"].isin(rs["patent"])]
ld = ld.loc[ld["patent"].isin(rs["patent"])]

re = dict(zip(re["patent_id"], re["examiner_id"]))
ld = dict(zip(ld["patent"], ld["lawyer_id"]))

# Top MSAs & Primary classes
pdf = fastparquet.ParquetFile("RawData/Cleaned/patent_loc_unique_us_0628.parq").to_pandas(["patent", "primclass", "inv_msa"])
dup_pats = pd.read_pickle("RawData/Cleaned/duplicate_pattext_0712.pkl").tolist()
# Get relevant US Patents
pdf = pdf.loc[~pdf["patent"].isin(dup_pats)]

topn["inv_msa"] = pdf["inv_msa"].value_counts()[:100].index.tolist()
topn["primclass"] = pdf["primclass"].value_counts()[:100].index.tolist()

rs["tp_inv_msa"] = rs["patent"].map(dict(zip(pdf["patent"], pdf["inv_msa"])))
del(pdf)

rs["tp_examiner"] = rs["patent"].map(re)
rs["tp_lawyer"] = rs["patent"].map(ld)
    
# Fixed effects
for c in ["primclass", "inv_msa", "examiner", "lawyer"]:
    rs["tp_{0}_FE".format(c)] = rs["tp_{0}".format(c)].astype(str)
    rs.loc[~(rs["tp_{0}".format(c)].isin(topn[c])), "tp_{0}_FE".format(c)] = "other"

    print(rs["tp_{0}_FE".format(c)].value_counts()[:10])
    # Drop original
    try:
        rs = rs.drop("tp_{0}".format(c), 1)
    except Exception:
        pass

rs["inv_msa_match"] = rs["inv_msa_match"].astype(bool)
rs.to_pickle("DataStore/2018-10/Reg1016/JTH_cite_reg_1026.pkl")

rs.sample(20)
