==> Reg1016/4a-CombinedResults-1016.ipynb <==
import scipy as sp
import numpy as np
import pandas as pd
import timeit
import re
import json
import pickle
import fastparquet
import os
os.chdir('/mnt/t48/bighomes-active/sfeng/patentdiffusion/')
seed = 3
import seaborn as sns
import matplotlib
import matplotlib.pyplot as plt
from sklearn.preprocessing import StandardScaler
scaler = StandardScaler()
import datetime
import statsmodels.api as sm
import statsmodels.formula.api as smf
import statsmodels.iolib.summary2 as summary2

pathdir = "DataStore/2018-10/Reg1016/"
reg_f = "reg_model_1016.pkl"

regs = pickle.load(open(pathdir+reg_f, "rb"))

# Model names to equations
regs["model_names_eqn"] = dict(zip(regs["model_names"].tolist(), regs["model_names"].index.tolist()))

# JTH Rep
jr = pd.read_csv("DataStore/2018-07-P3/JTHReg0727/JTH_res_out_0726.csv",index_col=0)

# FE results (0930 - uses full dataset) and Sim results (1003 - uses only those with sim pc, sim pc msa values)
rr = pd.read_csv("DataStore/2018-10/Reg1016/reg_pairs_out_HC1_1016.csv",index_col=0)

# Use reg_model_1002 to assign model numbers
rr["Model Num"] = rr["Model"].map(regs["model_names_eqn"])

# Get rid of models without a number (squared ones excluded from analysis)
print(len(rr))
rr = rr.loc[rr["Model Num"].notnull()]
print(len(rr))
rr["samp"] = rr["samp"].apply(eval)
rr = pd.concat([jr,rr],axis=0)
# Reset index
rr = rr.reset_index(drop=True)
# Integer model number
rr["Model Num"] = rr["Model Num"].astype(int)

rr["id"] = list(zip(rr["samp"], rr["Model Num"]))
rr["id"] = rr["id"].astype(str)

# KS dict
ksd = rr[["id", "LKS"]].drop_duplicates()
ksd = dict(zip(ksd["id"], ksd["LKS"]))

# Sample dict
sd = rr[["id", "samp"]].drop_duplicates()
sd = dict(zip(sd["id"], sd["samp"]))

rr.head()

### 1. Results with Year FE and PC Controls

repl1 = pickle.load(open("DataStore/2018-07-P3/reg_names_dict.pkl", "rb"))
# repl2 = {k: "\multicolumn{2}{c}{"+v+"}" if len(v)>1 else v for k,v in repl1.items() }

# Index by each table: Norm/Raw, KS, Sample
tab_ind = {}
tab_ind[("raw", "pct_cite", "JTH", "No FE")] = ("JTH Rep", 0)
tab_ind[("norm", "pct_cite", "JTH", "No FE")] = ("JTH Rep", 2)
tab_ind[("raw", "sim_cite", "JTH", "No FE")] = ("JTH Targ Sim", 4)
tab_ind[("norm", "sim_cite", "JTH", "No FE")] = ("JTH Targ Sim", 6)

tab_ind[("raw", "pct_cite", "JTH", "PC FE")] = ("JTH Rep", 1)
tab_ind[("norm", "pct_cite", "JTH", "PC FE")] = ("JTH Rep", 3)
tab_ind[("raw", "sim_cite", "JTH", "PC FE")] = ("JTH Targ Sim", 5)
tab_ind[("norm", "sim_cite", "JTH", "PC FE")] = ("JTH Targ Sim", 7)

for samp in ["naics_name", "primclass"]:
    if samp == "naics_name":
#         mods = ["PC FE-Year FE", "PC M-Year FE"]
        mods = ["PC FE-Year FE", "PC M-Year FE", "PC M", "CC M", "NPC M", "Inv M", "Lawyer M", "Examiner M"]
    else:
        mods = ["PC FE-Year FE", "Exam FE-Year FE"]
#     for dm in ["docvecs", "ldavecs", "tp_pct_common_cited"]:
    for dm in ["docvecs", "tp_pct_common_cited"]:
        for mod in mods:
#             tab_ind[("raw", dm, samp, mod)] = ((dm, samp), regs["model_names_eqn"][mod])
            tab_ind[("norm", dm, samp, mod)] = ((dm, samp), regs["model_names_eqn"]["N "+mod])

#### 1.1 Results by year

# Results by year
ygs = ["index", "1975-85", "1985-95", "1995-05", "2005-15"]
tabp = {}
for k in tab_ind.keys():
    
    i = str(tab_ind[k])
    try:
        mod = rr.loc[(rr["id"] == i), ["Model", "Model Num", "samp"]]\
        .drop_duplicates().values[0]
    except Exception:
        continue
    
    if "JTH" in k:
#         ygs = ["index", "1975-85", "1985-95", "1995-05",]
        ygs = ["index", "1975-85", "1985-95", "1995-05", "2005-15"]
    else:
        ygs = ["index", "1975-85", "1985-95", "1995-05", "2005-15"]
        
    # Select appropriate years
    tab = rr.loc[(rr["id"] == i), ygs].fillna("").set_index("index")

    # Add other columns
    if "No FE" in k:
        tab.loc["Year FE"] = [True]*len(tab.columns)
        tab.loc["PC FE"] = [False]*len(tab.columns)
    else:
        tab.loc["Year FE"] = [True]*len(tab.columns)
        tab.loc["PC FE"] = [True]*len(tab.columns)

#     # Pandas tables
    tab = tab.reset_index()
    # Replace "index"
    tab = tab.rename(columns={"index":""})
    # Replace everything else
    tab = tab.replace(repl1).copy()

    tabp[k] = tab
tab_f = tabp      

for k,v in tab_f.items():
    if (k[0] == "norm"):
        print("\n"+str(k)+"\n"+"\n")
        ncols = len(v.columns)
        with pd.option_context("max_colwidth", 1000):
            print(v.to_latex(index=False,escape=False, column_format="lccc"))

# for k,v in tab_f.items():
#     if (k[0] == "raw"):
#         print("\n"+str(k)+"\n"+"\n")
#         ncols = len(v.columns)
#         with pd.option_context("max_colwidth", 1000):
#             print(v.to_latex(index=False,escape=False, column_format="lccc"))

### 2. Results with Field Sim

repl1 = pickle.load(open("DataStore/2018-07-P3/reg_names_dict.pkl", "rb"))
dms = ["docvecs", "ldavecs", "tp_pct_common_cited"]
# dms = ["docvecs"]
# repl2 = {k: "\multicolumn{2}{c}{"+v+"}" if len(v)>1 else v for k,v in repl1.items() }
            
# Add names for standard errors
ygs = ["index", "1975-85", "1985-95", "1995-05", "2005-15"]
rr2 = rr.copy()
se_index = rr2[ygs+["index"]].loc[rr2["index"].isnull()].index
above_var = rr2.loc[se_index-1, "index"].tolist()
rr2.loc[se_index, "index"] = ["(s.e.) "+v for v in above_var]

#### 2.1 Results by year

# Index by each table: Norm/Raw, KS, Sample
tab_ind = {}
for samp in ["naics_name", "primclass"]:
    mods_all = ["PC M-Sim PC", "PC M-Sim PC MSA", "PC M-Int PC", "PC M-Int PC MSA",
                "Exam FE-Sim PC MSA", "Exam FE-Int PC MSA",
                "Exam FE-Break Int PC MSA"]
    if samp == "naics_name":
        mods = mods_all[:4]
    else:
        mods = mods_all[4:]
    for dm in dms:
        for mod in mods:
#             tab_ind[("raw", dm, samp, mod)] = ((dm, samp), regs["model_names_eqn"][mod])
            tab_ind[("norm", dm, samp, mod)] = ((dm, samp), regs["model_names_eqn"]["N "+mod])

tab_ind_yr = tab_ind

tab_no_c = {}
tab_c = {}
no_c_ind = ["C(common_est_inv)[T.1.0]", "C(common_pat_inv)[T.True]", 
            "C(lawyer_match)[T.True]", "C(examiner_match)[T.True]",
           "C(primclass_match)[T.True]"]
no_c_ind = no_c_ind+["(s.e.) "+c for c in no_c_ind]

for k in tab_ind_yr.keys():
    
    i = str(tab_ind_yr[k])
    
    try:
        mod = rr2.loc[(rr2["id"] == i), ["Model", "Model Num", "samp"]]\
    .drop_duplicates().values[0]
    except Exception:
        continue

    tab = rr2.loc[(rr2["id"] == i), ygs].fillna("").set_index("index")

    # Add other columns
    tab.loc["Year \& PC FE"] = [True]*len(tab.columns)
    tab.loc["MSA, Inv, Lawyer,  Examiner Match \& FE"] = [True]*len(tab.columns)

    # Create gaps
    tab = tab.fillna("").reset_index()

    #1. Create table without controls
    tab2 = tab.copy()
    tab2 = tab2.loc[~(tab2["index"].isin(no_c_ind))].reset_index(drop=True)
    # Get rid of se labels
    tab2.loc[tab2["index"].apply(lambda x: "(s.e.)" in x), "index"] = ""
    tab2 = tab2.rename(columns={"index":""})
    tab2 = tab2.replace(repl1).copy()

    tab_no_c[k] = tab2

    #2. Original tables with controls
    # Get rid of se labels
    tab.loc[tab["index"].apply(lambda x: "(s.e.)" in x), "index"] = ""
    tab = tab.rename(columns={"index":""})

    # Pandas tables
    tab = tab.replace(repl1).copy()
    tab_c[k] = tab
tab_c_yr = tab_c
tab_no_c_yr = tab_no_c

for k,v in tab_no_c_yr.items():
    if (k[0]=='norm') & (k[1]=='docvecs'):
        print(k)
        display(v)

for k,v in tab_no_c_yr.items():
    if (k[0]=='norm') & (k[1]=='docvecs'):
        print("\n"+str(k)+"\n"+"\n")
        ncols = len(v.columns)
        with pd.option_context("max_colwidth", 1000):
            print(v.to_latex(index=False,escape=False, column_format="lcccc"))

#### 2.2 Collected results - not doing this yet

# Index by each table: Norm/Raw, KS, Sample
tab_ind = {}
for samp in ["naics_name", "primclass"]:
    mods_all = ["PC M-Sim PC", "PC M-Sim PC MSA", "PC M-Int PC", "PC M-Int PC MSA",
                "Exam FE-Sim PC MSA", "Exam FE-Int PC MSA",
                "Exam FE-Break Int PC MSA"]
    if samp == "naics_name":
        mods = mods_all[:4]
    else:
        mods = mods_all[4:]
    mods_n = dict(zip(mods_all, range(1,len(mods_all)+1)))
#     for dm in ["docvecs", "ldavecs", "tp_pct_common_cited"]:
    for dm in dms:
        for mod in mods:
#             tab_ind[("raw", dm, samp, mod)] = ((dm, samp), regs["model_names_eqn"][mod])
            tab_ind[("norm", dm, samp, mod)] = ((dm, samp), regs["model_names_eqn"]["N "+mod])

tab_ind_collected = tab_ind

tab_no_c = {}
no_c_ind = ["C(common_est_inv)[T.1.0]", "C(common_pat_inv)[T.True]", "C(lawyer_match)[T.True]",
            "C(primclass_match)[T.True]", "C(examiner_match)[T.True]"]
no_c_ind = no_c_ind+["(s.e.) "+c for c in no_c_ind]

for k in tab_ind_collected.keys():
    
    i = str(tab_ind_collected[k])
    
    mod = rr2.loc[(rr2["id"] == i), ["Model", "Model Num", "samp"]]\
.drop_duplicates().values[0]

    tab = rr2.loc[(rr2["id"] == i), ygs].fillna("").set_index("index")

    # Add other columns
    tab.loc["Year FE"] = [True]*len(tab.columns)
    tab.loc["PC FE"] = [True]*len(tab.columns)

    if ("All FE" in k[3]) or ("Inv FE" in k[3]):
        tab.loc["Inv \& Lawyer Match"] = [True]*len(tab.columns)
    else:
        tab.loc["Inv \& Lawyer Match"] = [False]*len(tab.columns)

    # Create gaps
    tab = tab.fillna("").reset_index()

    #1. Create table without controls
    tab2 = tab.copy()
    tab2 = tab2.loc[~(tab2["index"].isin(no_c_ind))]
    # Set back index
    tab2 = tab2.set_index("index")
    
#     #
#     # Get rid of se labels
#     tab2.loc[tab2["index"].apply(lambda x: "(s.e.)" in x), "index"] = ""
#     tab2 = tab2.rename(columns={"index":""})
#     tab2 = tab2.replace(repl1).copy()

    tab_no_c[k] = tab2


tab_collected = tab_no_c

tab_col = {}
tab_col["norm"] = {}
# tab_col["raw"] = {}

keys = {}
keys["norm"] = [i for i in list(tab_collected.keys()) if "norm" in i]
# keys["raw"] = [i for i in list(tab_collected.keys()) if "raw" in i]
for m in keys.keys():
    for i,k in enumerate(keys[m]):
        # If group matches m, and is even
        if (m in k) & (i % 2) == 0:
            # Existing index order
            ind1 = tab_collected[keys[m][i+1]].index

            # First table
            a = tab_collected[keys[m][i]].copy()
            a.columns = pd.MultiIndex.from_product([["({0})".format(mods_n[keys[m][i][3]])], a.columns.tolist()])
            b = tab_collected[keys[m][i+1]].copy()
            b.columns = pd.MultiIndex.from_product([["({0})".format(mods_n[keys[m][i+1][3]])], b.columns.tolist()])
            tab2 = pd.concat([a, b], axis=1)
            tab2 = tab2.reindex(ind1)

            # Fill nan
            tab2 = tab2.fillna("").copy()

            # Get rid of se labels
            tab2.index = ["" if "(s.e.)" in x else x for x in tab2.index ]
            # Rename index
            tab2.index.name = "$KS="+repl1[k[1]][1::]
            # Get rid of se labels
            tab2 = tab2.reset_index(col_level=1)

            tab2 = tab2.replace(repl1).copy()

            tab_col[m][k] = tab2

m = "norm"
for k,v in tab_col[m].items():
    print(k)
    display(v)

m = "norm"
for k,v in tab_col[m].items():
    print("\n"+str(k)+"\n"+"\n")
    with pd.option_context("max_colwidth", 1000):
        print(v.to_latex(index=False,escape=False, column_format="l|cccc|cccc", multicolumn_format="c"))

m = "raw"
for k,v in tab_col[m].items():
    print("\n"+str(k)+"\n"+"\n")
    with pd.option_context("max_colwidth", 1000):
        print(v.to_latex(index=False,escape=False, column_format="l|cccc|cccc", multicolumn_format="c"))

#### 2.3 Localization estimate by model for norm, docvecs

local_est = pd.DataFrame()
for k,v in tab_no_c_yr.items():
    if (k[0] == "norm") & (k[1] == "docvecs"):
        v2 = v.loc[[0,1]]
        v2[""] = mods_n[k[3]]
        v2["var"] = ["est", "se"]
        v2["Model Name"] = k[3]
        local_est = local_est.append(v2, ignore_index=True)
        
local_est = local_est.sort_values(["", "var"]).reset_index(drop=True).drop("var",1)
local_est[""] = local_est[""].apply(lambda x: "({0})".format(x))
local_est.loc[list(range(1,len(local_est),2)), ""] = ""

local_est
