==> Reg1016/4a-CombinedResults-1026.ipynb <==
import scipy as sp
import numpy as np
import pandas as pd
import timeit
import re
import json
import pickle
import fastparquet
import os
os.chdir('/mnt/t48/bighomes-active/sfeng/patentdiffusion/')
seed = 3
import seaborn as sns
import matplotlib
import matplotlib.pyplot as plt
from sklearn.preprocessing import StandardScaler
scaler = StandardScaler()
import datetime
import statsmodels.api as sm
import statsmodels.formula.api as smf
import statsmodels.iolib.summary2 as summary2

Based on: https://sfengc7.stern.nyu.edu:8888/notebooks/patentdiffusion/201808Results/Reg1016/4a-CombinedResults-1016.ipynb, https://sfengc7.stern.nyu.edu:8888/notebooks/patentdiffusion/201808Results/Reg1016/4b-RegCoefficients-1016.ipynb

pathdir = "DataStore/2018-10/Reg1016/"
reg_f = "reg_model_1016.pkl"

regs = pickle.load(open(pathdir+reg_f, "rb"))

# Model names to equations
regs["model_names_eqn"] = dict(zip(regs["model_names"].tolist(), regs["model_names"].index.tolist()))

# JTH Rep
jr = pd.read_csv("DataStore/2018-10/Reg1016/JTH_cite_reg_1026.csv",index_col=0)
jr = jr.replace({"$I(MSA \, Match)$":"C(inv_msa_match)[T.True]",
                "Examiner FE-Year FE":"Exam FE-Year FE",
                "N Examiner FE-Year FE":"N Exam FE-Year FE"})
jr["samp"] = "JTH Rep"

# FE results (0930 - uses full dataset) and Sim results (1003 - uses only those with sim pc, sim pc msa values)
rr = pd.read_csv("DataStore/2018-10/Reg1016/reg_pairs_out_HC1_1016.csv",index_col=0)

# Use reg_model_1002 to assign model numbers
rr["Model Num"] = rr["Model"].map(regs["model_names_eqn"])

# Get rid of models without a number (squared ones excluded from analysis)
print(len(rr))
rr = rr.loc[rr["Model Num"].notnull()]
print(len(rr))
rr["samp"] = rr["samp"].apply(eval)
rr = pd.concat([jr,rr],axis=0)
# Reset index
rr = rr.reset_index(drop=True)
# Integer model number
rr["Model Num"] = rr["Model Num"].astype(int)
# Equation id
rr["id"] = ["({0}, {1}, {2})".format(i,j,k) for i,j,k in zip(rr["samp"], rr["LKS"], rr["Model"])]
rr["id"] = rr["id"].astype(str)

# Add S.E.
# S.E. index
se_ind = rr.loc[rr["index"].isnull()].index.tolist()

for ind in se_ind:
    rr.loc[ind, "index"] = "SE {0}".format(rr.loc[ind-1, "index"])

rr["Model"].unique()

### 1. Year and PC Controls

# Regression models
# Primclass
rm_pc = [
    # JTH norm percentage match PC FE
    "(JTH Rep, norm_perc_match_10, N PC FE-Year FE)",
    # pcc N PC FE
    "(('tp_pct_common_cited', 'primclass'), norm_tp_pct_common_cited, N PC FE-Year FE)",
    # docvecs N PC FE
    "(('docvecs', 'primclass'), norm_sim_docvecs, N PC FE-Year FE)",  
]

rm_pc_n = ["Citations Benchmark", "Pct Common Cited Pats", "Sim DocVecs, PC FE"]

# NAICS
rm_n = [
    # JTH norm percentage match PC FE
    "(JTH Rep, norm_perc_match_10, N PC FE-Year FE)",
    # pcc N PC FE
    "(('tp_pct_common_cited', 'naics_name'), norm_tp_pct_common_cited, N PC FE-Year FE)",
    # docvecs N PC FE
    "(('docvecs', 'naics_name'), norm_sim_docvecs, N PC FE-Year FE)", 
]

# rm_n_n = ["Citations Benchmark", "Pct Common Cited Pats", "Sim DocVecs, PC FE", "Sim DocVecs, Sim PC", "Sim DocVecs, Sim PC MSA"]
rm_n_n = rm_pc_n

rm_d = dict(zip(rm_pc, rm_pc_n))
rm_d.update(dict(zip(rm_n, rm_n_n)))

### NAICS Table

# Localization estimates
t = {}
for l in [("naics_name", rm_n), ("primclass", rm_pc)]:
    rc = rr.loc[rr["id"].isin(l[1]) & rr["index"].isin(["C(inv_msa_match)[T.True]",\
            "SE C(inv_msa_match)[T.True]", "$N$", "Adjusted $R^2$"])].copy()
    rc["id"] = rc["id"].map(rm_d)
    rc.loc[rc["index"].isin(["$N$", "Adjusted $R^2$"]), "id"] =\
    rc.loc[rc["index"].isin(["$N$", "Adjusted $R^2$"]), "index"]
    # Fixed effects columns
    rc["Year \& PC FE"] = ""
    rc["Other Controls"] = ""
    rc.loc[(rc["index"] == "SE C(inv_msa_match)[T.True]"), ['id']] = ""
    rc.loc[(rc["index"] == "C(inv_msa_match)[T.True]"), ["Year \& PC FE"]] = True
    rc.loc[(rc["index"] == "C(inv_msa_match)[T.True]"), ["Other Controls"]] = False
    # cols = ['id', '1975-85', '1985-95', '1995-05', '2005-15', "Year \& PC FE", "Other Controls"]
    cols = ['id', '1975-85', '1985-95', '1995-05', '2005-15']
    rc = rc[cols]
    rc = rc.rename(columns={"id": ""})
    t[l[0]] = rc
t1 = t

for k,v in t1.items():
    print(k)
    print(v.to_latex(index=False, escape=False, column_format="lcccc"))

### 2. All further controls

# Regression models
# Primclass
rm_pc = [
    # JTH norm percentage match PC FE
    "(JTH Rep, norm_perc_match_10, N Exam FE-Year FE)",
    # pcc N PC FE
    "(('tp_pct_common_cited', 'primclass'), norm_tp_pct_common_cited, N Exam FE-Year FE)",
    # docvecs N PC FE
    "(('docvecs', 'primclass'), norm_sim_docvecs, N Exam FE-Year FE)",  
]

rm_pc_n = ["Citations Benchmark", "Pct Common Cited Pats", "Sim DocVecs, PC FE"]

# NAICS
rm_n = [
    # JTH norm percentage match PC FE
    "(JTH Rep, norm_perc_match_10, N Exam FE-Year FE)",
    # pcc N PC FE
    "(('tp_pct_common_cited', 'naics_name'), norm_tp_pct_common_cited, N PC M-Year FE)",
    # docvecs N PC FE
    "(('docvecs', 'naics_name'), norm_sim_docvecs, N PC M-Year FE)", 
]

# rm_n_n = ["Citations Benchmark", "Pct Common Cited Pats", "Sim DocVecs, PC FE", "Sim DocVecs, Sim PC", "Sim DocVecs, Sim PC MSA"]
rm_n_n = rm_pc_n

rm_d = dict(zip(rm_pc, rm_pc_n))
rm_d.update(dict(zip(rm_n, rm_n_n)))

# Localization estimates
t = {}
for l in [("naics_name", rm_n), ("primclass", rm_pc)]:
    rc = rr.loc[rr["id"].isin(l[1]) & rr["index"].isin(["C(inv_msa_match)[T.True]",\
            "SE C(inv_msa_match)[T.True]", "$N$", "Adjusted $R^2$"])].copy()
    rc["id"] = rc["id"].map(rm_d)
    rc.loc[rc["index"].isin(["$N$", "Adjusted $R^2$"]), "id"] =\
    rc.loc[rc["index"].isin(["$N$", "Adjusted $R^2$"]), "index"]
    # Fixed effects columns
    rc["Year \& PC FE"] = ""
    rc["Other Controls"] = ""
    rc.loc[(rc["index"] == "SE C(inv_msa_match)[T.True]"), ['id']] = ""
    rc.loc[(rc["index"] == "C(inv_msa_match)[T.True]"), ["Year \& PC FE"]] = True
    rc.loc[(rc["index"] == "C(inv_msa_match)[T.True]"), ["Other Controls"]] = False
    # cols = ['id', '1975-85', '1985-95', '1995-05', '2005-15', "Year \& PC FE", "Other Controls"]
    cols = ['id', '1975-85', '1985-95', '1995-05', '2005-15']
    rc = rc[cols]
    rc = rc.rename(columns={"id": ""})
    t[l[0]] = rc
t2 = t

for k,v in t2.items():
    print(k)
    print(v.to_latex(index=False, escape=False, column_format="lcccc"))
